Able to Add vehicles to garage and check if it's full.

+++ Garage_JoakimMalmstrom/Boat.cs
@@ -2,8 +2,8 @@
 {
     internal class Boat : Vehicle
     {
-        public float Length { get; set; }
-        public Boat(string regNumber, string color, int numWheels, float length) : base(regNumber, color, numWheels)
+        public double Length { get; set; }
+        public Boat(string regNumber, string color, int numWheels, double length) : base(regNumber, color, numWheels)
         {
             Length = length;
         }
